+++ src/tests/Q_125/ValidPalindrome_1Test.java
@@ -1,6 +1,6 @@
 package tests.Q_125;
 
-import leetcode.Lc125_ValidPalindrome_1;
+import leetcode.normal.Q_125.ValidPalindrome_1;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
@@ -9,7 +9,7 @@ public class ValidPalindrome_1Test {
 
     @Test
     public void isPalindrome() {
-        Lc125_ValidPalindrome_1 solution = new Lc125_ValidPalindrome_1();
+        ValidPalindrome_1 solution = new ValidPalindrome_1();
         assertTrue(solution.isPalindrome(""));
         assertTrue(solution.isPalindrome("A man, a plan, a canal: Panama"));
         assertFalse(solution.isPalindrome("race a car"));
