+++ src/tests/Q_760/FindAnagramMappingTest.java
@@ -1,6 +1,6 @@
 package tests.Q_760;
 
-import leetcode.Lc760_FindAnagramMapping;
+import leetcode.normal.Q_760.FindAnagramMapping;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
@@ -12,7 +12,7 @@ public class FindAnagramMappingTest {
         int[] A = {12, 28, 46, 32, 50};
         int[] B = {50, 12, 32, 46, 28};
 
-        Lc760_FindAnagramMapping solution = new Lc760_FindAnagramMapping();
+        FindAnagramMapping solution = new FindAnagramMapping();
 
         assertArrayEquals(solution.anagramMappings(A, B), new int[]{1, 4, 3, 2, 0});
