+++ src/tests/Q_819/MostCommonWordTest.java
@@ -1,6 +1,6 @@
 package tests.Q_819;
 
-import leetcode.Lc819_mostCommonWord;
+import leetcode.normal.Q_819.MostCommonWord;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
@@ -9,7 +9,7 @@ public class MostCommonWordTest {
 
     @Test
     public void mostCommonWord() {
-        Lc819_mostCommonWord solution = new Lc819_mostCommonWord();
+        MostCommonWord solution = new MostCommonWord();
         String para = "Bob. hIt, balL";
         String[] banned = {"bob", "hit"};
         assertEquals(solution.mostCommonWord(para, banned), "ball");
